+++ modules/tinymce/src/core/test/ts/browser/newline/InsertNewLineTest.ts
@@ -1,5 +1,6 @@
 import { ApproxStructure } from '@ephox/agar';
 import { after, before, context, describe, it } from '@ephox/bedrock-client';
+import { Type } from '@ephox/katamari';
 import { TinyAssertions, TinyHooks, TinySelections, TinyState } from '@ephox/wrap-mcagar';
 
 import Editor from 'tinymce/core/api/Editor';
@@ -16,6 +17,21 @@ describe('browser.tinymce.core.newline.InsertNewLineTest', () => {
 
   const insertNewline = (editor: Editor, args: Partial<EditorEvent<KeyboardEvent>>) => {
     InsertNewLine.insert(editor, args as EditorEvent<KeyboardEvent>);
+  };
+
+  const setSelectionToBody = (editor: Editor, offset: number) => {
+    const sel = editor.selection.getSel();
+
+    if (Type.isNullable(sel)) {
+      throw new Error();
+    }
+
+    const range = editor.contentDocument.createRange();
+    range.setStart(editor.dom.getRoot(), offset);
+    range.setEnd(editor.dom.getRoot(), offset);
+
+    sel.removeAllRanges();
+    sel.addRange(range);
   };
 
   context('Enter in paragraph', () => {
@@ -141,6 +157,33 @@ describe('browser.tinymce.core.newline.InsertNewLineTest', () => {
         insertNewline(editor, { });
         TinyAssertions.assertContent(editor, '<div contenteditable="false"><div contenteditable="true"><p>a</p><p>b</p></div></div>');
       });
+    });
+
+    it('TINY-9813: Placed a cursor is placed after a table, with a noneditable afterwards', () => {
+      const editor = hook.editor();
+      editor.setContent('<table><tbody><tr><td><br></td></tr></tbody></table><div contenteditable="false"></div>');
+      setSelectionToBody(editor, 1);
+      insertNewline(editor, { });
+      TinyAssertions.assertContent(editor, '<table><tbody><tr><td>&nbsp;</td></tr></tbody></table><p>&nbsp;</p><div contenteditable="false">&nbsp;</div>');
+      TinyAssertions.assertCursor(editor, [ 1 ], 0);
+    });
+
+    it('TINY-9813: Placed a cursor is placed after a table, with an editable afterwards', () => {
+      const editor = hook.editor();
+      editor.setContent('<table><tbody><tr><td><br></td></tr></tbody></table><div contenteditable="true">&nbsp;</div>');
+      setSelectionToBody(editor, 1);
+      insertNewline(editor, { });
+      TinyAssertions.assertContent(editor, '<table><tbody><tr><td>&nbsp;</td></tr></tbody></table><div contenteditable="true">&nbsp;</div><div contenteditable="true">&nbsp;</div>');
+      TinyAssertions.assertCursor(editor, [ 2 ], 0);
+    });
+
+    it('TINY-9813: Placed a cursor is placed after a table, with nothing', () => {
+      const editor = hook.editor();
+      editor.setContent('<table><tbody><tr><td><br></td></tr></tbody></table>');
+      setSelectionToBody(editor, 1);
+      insertNewline(editor, { });
+      TinyAssertions.assertContent(editor, '<table><tbody><tr><td>&nbsp;</td></tr></tbody></table><p>&nbsp;</p>');
+      TinyAssertions.assertCursor(editor, [ 1 ], 0);
     });
   });